Unify import pages
